Log write added in category

+++ src/InventoryManagementSystem/Inventory.Presentation/Controllers/CategoryController.cs
@@ -10,19 +10,23 @@ namespace Inventory.Presentation.Controllers
     [Authorize]
     public class CategoryController : Controller
     {
+        private readonly ILogger<CategoryController> _logger;
         private readonly ICategoryManagementService _categoryManagementService;
-        public CategoryController(ICategoryManagementService categoryManagementService)
+        public CategoryController(ILogger<CategoryController> logger, ICategoryManagementService categoryManagementService)
         {
+            _logger = logger;
             _categoryManagementService = categoryManagementService;
         }
         public async Task<IActionResult> Index()
         {
             var categories = await _categoryManagementService.GetAllCategories();
+            _logger.LogInformation("Category Index loaded");
             return View(categories);
         }
 
         public IActionResult Create()
         {
+            _logger.LogInformation("Category Create view page loaded");
             return View();
         }
 
@@ -37,15 +41,21 @@ namespace Inventory.Presentation.Controllers
                     if(model.Name == null)
                         return View(model);
                     await _categoryManagementService.CreateCategory(model.Name, model.Description);
-                    TempData["success"] = "Category created successfully";
+
+                    TempData["Success"] = "Category created successfully";
+                    _logger.LogInformation("New Category created....");
+
                     return RedirectToAction("Index");
                 }
                 catch (Exception ex)
                 {
-                    Console.WriteLine(ex.Message);
+                    TempData["Error"] = "Failed to create category";
+                    _logger.LogInformation("Error occured in create category post action");
+                    _logger.LogError($"Error: {ex}");
                 }
             }
-            TempData["error"] = "Error occured";
+            _logger.LogInformation("Error occured for create category model validation failed..");
+            TempData["Error"] = "Failed to create category";
 
             return View(model);
         }
@@ -55,12 +65,16 @@ namespace Inventory.Presentation.Controllers
         {
             var category = await _categoryManagementService.GetCategoryIdAsync(id);
             if (category == null)
+            {
                 throw new Exception("Category not found");
+            }
 
             var model = new UpdateCategoryModel();
             model.Description = category.Description;
             model.Name = category.Name;
             model.Id = category.Id;
+
+            _logger.LogInformation("Category update view page loaded...");
 
             return View(model);
         }
@@ -75,16 +89,22 @@ namespace Inventory.Presentation.Controllers
                 {
                     if (model.Name == null)
                         return View(model);
+
                     await _categoryManagementService.UpdateCategoryAsync(model.Id, model.Name, model.Description);
                     TempData["success"] = "Category updated successfully";
+                    _logger.LogInformation("Category  update successfully....");
+
                     return RedirectToAction("Index");
                 }
                 catch (Exception ex)
                 {
-                    Console.WriteLine(ex.Message);
+                    TempData["Error"] = "Failed to update category";
+                    _logger.LogInformation("Error occured in update category post action");
+                    _logger.LogError($"Error: {ex}");
                 }
             }
-            TempData["error"] = "Error occured";
+            TempData["Error"] = "Failed to update category";
+            _logger.LogInformation("Error occured for update category model validation failed..");
 
             return View(model);
         }
@@ -100,26 +120,42 @@ namespace Inventory.Presentation.Controllers
             model.Name = category.Name;
             model.Id = category.Id;
 
+            _logger.LogInformation("Category delete view page loaded...");
+
+
             return View(model);
         }
 
         [HttpPost]
         public async Task<IActionResult> Delete(UpdateCategoryModel model)
         {
-            try
+            if (ModelState.IsValid)
             {
-                var category = await _categoryManagementService.GetCategoryIdAsync(model.Id);
-                if (category == null)
-                    throw new Exception("Category not found");
+                try
+                {
+                    var category = await _categoryManagementService.GetCategoryIdAsync(model.Id);
+                    if (category == null)
+                        throw new Exception("Category not found");
 
-                await _categoryManagementService.RemoveCategoryAsync(category);
-                return RedirectToAction("Index");
+                    await _categoryManagementService.RemoveCategoryAsync(category);
+
+                    TempData["success"] = "Category deleted successfully";
+                    _logger.LogInformation("Category  deleted successfully....");
+
+                    return RedirectToAction("Index");
+                }
+                catch (Exception ex)
+                {
+                    TempData["Error"] = "Failed to delete category";
+                    _logger.LogInformation("Error occured in delete category post action");
+                    _logger.LogError($"Error: {ex}");
+                }
             }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
-            }
-            return View();
+
+            TempData["Error"] = "Failed to delete category";
+            _logger.LogInformation("Error occured for delete category model validation failed..");
+
+            return View(model);
         }
     }
 }
